refactor: split assignmentDefinitionPartials.js into validation + transport
- Create assignmentDefinitionValidation.js (707 lines, 20 validation functions)
- Create assignmentDefinitionTransport.js (220 lines, 6 handler functions)
- Delete original assignmentDefinitionPartials.js (916 lines)
- Update setupGlobals.js to load both new modules
- Update all test imports for dual-file structure
- Section 5 complete in ACTION_PLAN_DATE_HARDENING.md
